Move index builder into separate file
